Moved 'button' into correct component folder
