+++ src/screens/EscolhaCapituloScreen.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components/native";
-import Button from "../../components/Button";
+import Button from "../components/Button";
 import NavBar from "../components/NavBar";
 import { ArrowFatLeft } from "phosphor-react-native";
 import ARC from "../../assets/ARC.json";
